chore(platforms, iam): move redis cache to platforms
